# Optimize `Hash#transform_{keys,values}`