@straight-shoota 2024-04-18
Initializing the transformed has with the current size seems like a good idea. We could also consider using the current capacity in order to make it a copy with identical properties, just transformed values/keys.

But I'm not sure about optimizing the capacity for small hashes. The default minimum capacity for a hash has its reasons. It's not highly optimized for the use case of very small hashes, but that's that common and the overhead is minimal.
It will overall be less efficient if more items are added afterwards. I suppose an argument could be made that transformed hashes are less likely to get new additions. But that's speculative and surely there are many use cases where it doesn't hold.

So my suggestion would be to drop the special handling for small hashes.