@ysbaddaden 2024-04-19
Let's do some archeology:

In #8017 `Hash#initialize` called `#malloc_entries(initial_capacity / 2)` which correlated with the other `#malloc_entries(4)` since initial capacity is supposed to be 8.

But #8182 patched `#initialize` to now call `#malloc_entries(initial_capacity)` but it didn't change the other `#malloc_entries(4)` and now they don't correlate anymore.

Said differently: magic numbers :face_exhaling: 

All that to say that magic numbers really are as bad as we say they are + I agree with @straight-shoota's statement above: let's just always use the capacity, as other calls to `#malloc_entries` do.